+++ src/snowflake/cli/_plugins/notebook/commands.py
@@ -13,14 +13,33 @@
 # limitations under the License.
 
 import logging
+from typing import Dict
 
 import typer
+from click import UsageError
 from snowflake.cli._plugins.notebook.manager import NotebookManager
+from snowflake.cli._plugins.notebook.notebook_entity_model import NotebookEntityModel
 from snowflake.cli._plugins.notebook.types import NotebookStagePath
-from snowflake.cli.api.commands.flags import identifier_argument
+from snowflake.cli._plugins.workspace.manager import WorkspaceManager
+from snowflake.cli.api.cli_global_context import get_cli_context
+from snowflake.cli.api.commands.decorators import (
+    with_project_definition,
+)
+from snowflake.cli.api.commands.flags import (
+    IfNotExistsOption,
+    ReplaceOption,
+    entity_argument,
+    identifier_argument,
+)
 from snowflake.cli.api.commands.snow_typer import SnowTyperFactory
+from snowflake.cli.api.entities.common import EntityActions
+from snowflake.cli.api.exceptions import NoProjectDefinitionError
 from snowflake.cli.api.identifiers import FQN
-from snowflake.cli.api.output.types import MessageResult
+from snowflake.cli.api.output.types import (
+    CollectionResult,
+    CommandResult,
+    MessageResult,
+)
 from typing_extensions import Annotated
 
 app = SnowTyperFactory(
@@ -84,3 +103,52 @@ def create(
         notebook_file=notebook_file,
     )
     return MessageResult(message=notebook_url)
+
+
+@app.command(requires_connection=True)
+@with_project_definition()
+def deploy(
+    entity_id=entity_argument("notebook"),
+    replace: bool = ReplaceOption(
+        help="Replace existing Notebook if it already exists.",
+    ),
+    if_not_exists: bool = IfNotExistsOption(help="Skip if Notebook already exists."),
+    **options,
+) -> CommandResult:
+    """Uploads a notebook to a stage and creates a notebook. If entity_id is not provided,
+    deploys all notebooks defined in the project definition."""
+    cli_context = get_cli_context()
+    pd = cli_context.project_definition
+    if not pd.meets_version_requirement("2"):
+        raise UsageError(
+            "This command requires project definition of version at least 2."
+        )
+
+    notebooks: Dict[str, NotebookEntityModel] = pd.get_entities_by_type(
+        entity_type="notebook"
+    )
+    if not notebooks:
+        raise NoProjectDefinitionError(
+            project_type="notebook", project_root=cli_context.project_root
+        )
+    if entity_id:
+        if entity_id not in notebooks:
+            raise UsageError(f"No '{entity_id}' notebook in project definition file.")
+        notebooks = {entity_id: notebooks[entity_id]}
+
+    ws = WorkspaceManager(
+        project_definition=cli_context.project_definition,
+        project_root=cli_context.project_root,
+    )
+
+    deploy_results = []
+    for entity_id in sorted(notebooks):
+        result = ws.perform_action(
+            entity_id,
+            EntityActions.DEPLOY,
+            replace=replace,
+            if_not_exists=if_not_exists,
+        )
+        deploy_results.append(result)
+
+    return CollectionResult(deploy_results)