+++ src/snowflake/cli/api/project/schemas/entities/entities.py
@@ -24,6 +24,8 @@ from snowflake.cli._plugins.nativeapp.entities.application_package import (
     ApplicationPackageEntity,
     ApplicationPackageEntityModel,
 )
+from snowflake.cli._plugins.notebook.notebook_entity import NotebookEntity
+from snowflake.cli._plugins.notebook.notebook_entity_model import NotebookEntityModel
 from snowflake.cli._plugins.snowpark.snowpark_entity import (
     FunctionEntity,
     ProcedureEntity,
@@ -50,6 +52,7 @@ Entity = Union[
     ProcedureEntity,
     FunctionEntity,
     ComputePoolEntity,
+    NotebookEntity,
 ]
 EntityModel = Union[
     ApplicationEntityModel,
@@ -58,6 +61,7 @@ EntityModel = Union[
     FunctionEntityModel,
     ProcedureEntityModel,
     ComputePoolEntityModel,
+    NotebookEntityModel,
 ]
 
 ALL_ENTITIES: List[Entity] = [*get_args(Entity)]