+++ RELEASE-NOTES.md
@@ -49,6 +49,7 @@
 * Added support for glob pattern (except `**`) in artifact paths in snowflake.yml for Snowpark, requires ENABLE_SNOWPARK_GLOB_SUPPORT feature flag.
 * Added `--replace` flag to `snow spcs compute-pool create` command.
 * Added command `snow spcs compute-pool deploy`.
+* Add `snow notebook deploy` command, which bases on new notebook project definition.
 
 ## Fixes and improvements